Extract menu item fetching to hook
